Updated user-tab. CSS
Added additional foundation icons beside the user-tab links. I'm
becoming a fan of these icons really fast. Also, I adjusted the bottom
padding on the list (ul) of the user-tab links. It doesn't have any now,
so the user actually has to mouse over the name/picture area for the
drop-down, which seems to flow better.