CSS. Message list update.
Re-arranged the message-list panel a bit. Now a larger mail icon appears
in blue, coupled with the importance value. The user's gravatar is also
displayed beside the user's name. We decided for now to leave out
message timestamp in the list, but if needed it is just an <i> tag in
the <address> tag of the user.