CSS. Linking Inbox and Compose.
I edited the format of the Inbox and Compose so their "sidebars" were
similar in format. The design of the sidebar will most likely change
later on, but for now they should be consistent between the inbox and
compose. I also changed some of the formatting on the compose to tidy up
the Subject Line and the Message textareas (mainly spacing and padding).